Flytter data fetching fra UtbetalingerPanel til …Nye
